+++ changelog.md
@@ -1,1 +1,3 @@
 - fixed an issue on Mac
+- ported stuff from 1.21.1
+- packed packs compat